MyDNS の更新タスクを追加
inet-jp.info へアクセスしての IP addr 取得コードは Obsolete をつけて残しておきます
